Prepravljene i dodate DTO klase, kontroler za rezervacije...

+++ src/main/java/com/msmisa/TicketApp/resources/AbstractController.java
@@ -6,11 +6,14 @@ import org.apache.commons.logging.Log;
 import org.apache.commons.logging.LogFactory;
 import org.modelmapper.ModelMapper;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
 
+import com.msmisa.TicketApp.dao.DaoException;
 import com.msmisa.TicketApp.dao.GenericDao;
 
 public abstract class AbstractController<Entity, Key> {
@@ -35,16 +38,22 @@ public abstract class AbstractController<Entity, Key> {
 		return dao;
 	}
 
-
-
+	
 	@RequestMapping(value="/delete/{id}",
 			method=RequestMethod.DELETE)
-	public void delete(@PathVariable(value="id") Key id){
-		dao.delete(id);
+	public ResponseEntity<?> delete(@PathVariable(value="id") Key id){
+		try {
+			Entity e = dao.get(id);
+			if(e == null) {
+				return new ResponseEntity<String>("Entity doesn't exist.", HttpStatus.NO_CONTENT);
+			}
+			dao.delete(id);
+			return new ResponseEntity<String>("Success deleting entity", HttpStatus.OK);
+		} catch(DaoException e) {
+			e.printStackTrace();
+			return new ResponseEntity<String>("Error deleting entity", HttpStatus.NOT_ACCEPTABLE);
+		}
 	}
-
-	
-
 }
